pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "numerous-apps"
version = "0.4.3"
description = "Application package for numerous"
authors = [
    {name = "Lasse Thomsen", email = "lasse.thomsen@numerous.com"},
]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "jinja2==3.1.4",
    "anywidget==0.9.13",
    "numpy>=2.0.0",
    "pydantic==2.10.3",
    "PyJWT>=2.8.0",
]

readme = "docs/README.md"
license = {text = "MIT"}  # Adjust license as needed
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers"
]

[project.optional-dependencies]
# Authentication with database backend
auth-db = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "bcrypt>=4.1.0",
]

dev = [
    "anywidget[dev]==0.9.13",
    "python-dotenv==1.0.1",
    "numerous-widgets",
    "numpy>=2.0.0",
    "pytest==8.3.4",
    "ruff==0.6.4",
    "mypy==1.11.2",
    "coverage==7.6.1",
    "twine==5.1.1",
    "mkdocs==1.6.0",
    "mkdocstrings==0.26.2",
    "mkdocstrings-python==1.12.2",
    "mkdocs_snippet_plugin==1.0.2",
    "pre-commit==4.0.1",
    "black==24.10.0",
    "mkdocs-material==9.5.46",
    "mkdocs-gen-files==0.5.0",
    "mkdocs-section-index==0.3.9",
    "mkdocs-literate-nav==0.6.1",
    "marimo==0.10.7",
    "panel==1.5.5",
    "pytest-cov==4.1.0",
    "httpx==0.28.1",
    "pytest-asyncio==0.25.2",
    "playwright==1.49.1",
    "psutil==6.1.1",
]

[tool.poetry.dependencies]
python-dotenv = "^1.0.0"

[tool.semantic_release]
branch = "main"
commit_parser = "angular"
version_toml = ["pyproject.toml:project.version"]
template_dir = "templates"

[tool.semantic_release.remote.token]
env = "GH_TOKEN"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    '''chore(?:\([^)]*?\))?: .+''',
    '''ci(?:\([^)]*?\))?: .+''',
    '''refactor(?:\([^)]*?\))?: .+''',
    '''style(?:\([^)]*?\))?: .+''',
    '''test(?:\([^)]*?\))?: .+''',
    '''build\((?!deps\): .+)''',
    '''Merged? .*''',
    '''initial commit''',
]


[project.urls]
Homepage = "https://github.com/numerous-com/numerous-app"
Repository = "https://github.com/numerous-com/numerous-app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["numerous.*"]
namespaces = true

[tool.setuptools.package-data]
"numerous.apps" = [
    "py.typed",
    "js/*.js",
    "static/*.md",
    "templates/*.html.j2",
    "bootstrap_app/*",
    "bootstrap_app/static/*",
    "bootstrap_app/static/css/*",
    "bootstrap_app/widgets/*",
    "auth/**/*.py",
]

[project.scripts]
numerous-bootstrap = "numerous.apps.bootstrap:main"

[tool.ruff]
src = ["src"]
exclude = ["examples", "tests", "e-e-test"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN101",
    "D101",
    "D103",
    "D107",
    "D203",
    "D211",
    "D212",
    "FA100",
    "FA102",
    "ISC001",
    "COM812",
    "FBT001",
    "FBT002",
    "PLR0913",
    "G004",
    "EM101",
    "TRY003",
    "DTZ005",
    "EM102",
    "PLC0414",
]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.extend-per-file-ignores]
"tests/**" = ["INP001", "S101", "D100", "D103"]
# FastAPI patterns require Depends() in defaults and Any for protocol-based providers
# TID252: Relative imports are preferred within packages
# ARG001: Unused args used for dependency injection
# BLE001: Broad exception catches are needed for token validation
# D401: Docstring imperative mood not always applicable
# E402: Lazy imports are needed to avoid circular imports
# N802: Property names match class names
# SIM103, TRY300: Simplification suggestions not always better for readability
# S110: Silent exception passes for optional auth
# ANN204: __init__ return type annotation not needed
"src/numerous/apps/auth/**" = [
    "ANN204", "ANN401", "ARG001", "B008", "BLE001", "D401", "E402",
    "N802", "S110", "SIM103", "TID252", "TRY300"
]
"src/numerous/apps/app_server.py" = ["ANN401", "D401", "E402"]

[tool.mypy]
ignore_missing_imports = true
exclude = ["examples", "e-e-test"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=numerous.apps --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]

[tool.coverage.run]
source = ["numerous.apps"]
branch = false

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
ignore_errors = true
omit = [
    "tests/*",
    "setup.py",
]
